add creating account with version 2 of API

+++ app/src/main/java/com/ithirteeng/secondpatternsclientproject/domain/accounts/model/bank/CurrencyCode.kt
@@ -4,5 +4,5 @@ package com.ithirteeng.secondpatternsclientproject.domain.accounts.model.bank
 data class CurrencyCode(
     val code: String,
     val numericCode: Int,
-    val exchangeRate: Int,
+    val exchangeRate: Double,
 )
